+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2025-01-14
+
+### Added
+
+- Added description property to data journey flows.
+
 ## [0.4.1] - 2024-11-28
 
 ### Fixed